requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
bcrypt==4.2.0
PyMySQL==1.1.1
cryptography==43.0.1
python-dotenv==1.0.1
gunicorn==23.0.0

